requirements.txt
dependency-injector
Flask
Flask-SQLAlchemy
psycopg2-binary
pytest
pytest-cov
requests
